implemented learning_loop_create_date function and hacked together a solution in main
learning_loop_create_data in QLearning.py returns the list of average scores at the end of the learning loop instead of graphing the scores.

I implemented the graphing of the average scores of multiple agents using this new function. It is messy but I will create actual functions and clean up the code when I get the chance.